fix(animations): ensure consistent transition namespace ordering
When including a component in a template, the component's host element is immediately appended as
child of the parent node upon creation. Hence, `hostElement.parentNode` will be a valid reference.
However, if the parent component is being inserted as an embedded view---through `ngIf` for
example---then the parent's node itself will not have been insterted yet. This means that we cannot
properly determine the position of the transition namespace, as any `containsElement` check will
return false given that the partial DOM tree has not been inserted yet, even though it will be
contained within an existing transition namespace once the partial tree is attached.